# Migrate to new reflection API