--- a/src/test/php/inject/unittest/NewInstanceTest.class.php
+++ b/src/test/php/inject/unittest/NewInstanceTest.class.php
@@ -14,11 +14,11 @@ class NewInstanceTest {
    * Creates a unique and new fixture subclass with the given definition
    *
    * @param  [:var] $definition
-   * @return inject.unittest.fixture.Storage
+   * @return inject.unittest.fixture.Fixture
    */
   protected function newFixture($definition) {
     return ClassLoader::defineClass(
-      'inject.unittest.fixture.T'.uniqid(),
+      'inject.NewInstanceTest_'.uniqid(),
       'inject.unittest.fixture.Fixture',
       [],
       $definition
@@ -50,7 +50,7 @@ class NewInstanceTest {
   public function newInstance_performs_injection() {
     $inject= (new Injector())->bind(Storage::class, $this->storage);
     $fixture= $this->newFixture([
-      '#[\inject\Inject] __construct' => function(Storage $param) { $this->injected= $param; }
+      '#[Inject] __construct' => function(Storage $param) { $this->injected= $param; }
     ]);
 
     Assert::equals($this->storage, $inject->newInstance($fixture)->injected);
@@ -60,7 +60,7 @@ class NewInstanceTest {
   public function newInstance_performs_named_injection_using_array_form() {
     $inject= (new Injector())->bind(Storage::class, $this->storage, 'test');
     $fixture= $this->newFixture([
-      '#[\inject\Inject(name: "test")] __construct' => function(Storage $param) { $this->injected= $param; }
+      '#[Inject(name: "test")] __construct' => function(Storage $param) { $this->injected= $param; }
     ]);
 
     Assert::equals($this->storage, $inject->newInstance($fixture)->injected);
@@ -70,7 +70,7 @@ class NewInstanceTest {
   public function newInstance_performs_named_injection_using_string_form() {
     $inject= (new Injector())->bind(Storage::class, $this->storage, 'test');
     $fixture= $this->newFixture([
-      '#[\inject\Inject("test")] __construct' => function(Storage $param) { $this->injected= $param; }
+      '#[Inject("test")] __construct' => function(Storage $param) { $this->injected= $param; }
     ]);
 
     Assert::equals($this->storage, $inject->newInstance($fixture)->injected);
@@ -112,7 +112,7 @@ class NewInstanceTest {
     $inject= new Injector();
     $inject->bind(Storage::class, $this->storage);
     $fixture= $this->newFixture([
-      '#[\inject\Inject] __construct' => function(Storage $param, $verify= true) { $this->injected= [$param, $verify]; }
+      '#[Inject] __construct' => function(Storage $param, $verify= true) { $this->injected= [$param, $verify]; }
     ]);
 
     Assert::equals([$this->storage, true], $inject->newInstance($fixture)->injected);
@@ -121,7 +121,7 @@ class NewInstanceTest {
   #[Test, Expect(class: CannotInstantiate::class, message: '/Cannot instantiate .+/')]
   public function newInstance_catches_cannot_instantiate_when_creating_class_instances() {
     $this->newInstance(new Injector(), $this->newFixture('{
-      #[\inject\Inject]
+      #[Inject]
       private function __construct() { }
     }'));
   }
@@ -129,7 +129,7 @@ class NewInstanceTest {
   #[Test, Expect(class: ProvisionException::class, message: '/No bound value for type string named "endpoint"/')]
   public function newInstance_throws_when_value_for_required_parameter_not_found() {
     $this->newInstance(new Injector(), $this->newFixture('{
-      #[\inject\Inject(type: "string", name: "endpoint")]
+      #[Inject(type: "string", name: "endpoint")]
       public function __construct($uri) { }
     }'));
   }