See https://github.com/xp-framework/rfc/issues/338

## Performance

Before:

```bash
$ xp test src/test/php/
# ...
Test cases:  177 succeeded
Memory used: 3676.30 kB (3730.42 kB peak)
Time taken:  0.031 seconds (0.096 seconds overall)
```
After:

```bash
$ xp test src/test/php/
# ...
Test cases:  177 succeeded
Memory used: 3621.24 kB (3675.37 kB peak)
Time taken:  0.034 seconds (0.107 seconds overall)
```